+++ metadata.rb
@@ -4,4 +4,4 @@ maintainer_email 'phnahes@gmail.com'
 license          'All rights reserved'
 description      'Installs/Configures deploy'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.2'
+version          '0.1.7'