+++ CHANGELOG.md
@@ -3,6 +3,23 @@ deploy CHANGELOG
 
 This file is used to list changes made in each version of the deploy cookbook.
 
+0.1.7
+-----
+- [Paulo Nahes] - Bug Fix and add support for ubuntu and centos init scripts
+
+
+0.1.6
+-----
+- [Paulo Nahes] - Add resources to support too many environments automatically
+
+0.1.5
+-----
+- [Paulo Alves] - Improvements on exec commands
+
+0.1.4
+-----
+- [Paulo Nahes] - Bug Fix on Package Install 
+
 0.1.3
 -----
 - [Paulo Nahes] - Rewrited Cookbook, removed Providers, simplified without it.